requirements.txt
openai 
streamlit
python-dotenv
deep_translator
lingua-language-detector